Update scenario-sentiment-analysis-deep-learning.md
The original ModelingAndEvaluation.md is not actually linked to the file related to execute SentimentExtraction.py. 
The original ModelingAndEvaluationDocker is actually a guide on model details and modeling with CNTK as Backend.
Be good to specify the difference between executing on local and docker compute context.
Now, I modify the names of the markdown files and corresponding python files.
See if this is better.